+++ src/main/kotlin/team/themoment/datagsm/global/security/config/SecurityConfig.kt
@@ -14,6 +14,7 @@ import org.springframework.security.web.SecurityFilterChain
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
 import org.springframework.web.cors.CorsConfigurationSource
 import team.themoment.datagsm.global.config.DomainAuthorizationConfig
+import team.themoment.datagsm.global.security.handler.CustomAuthenticationEntryPoint
 import team.themoment.datagsm.global.security.jwt.JwtProvider
 import team.themoment.datagsm.global.security.jwt.filter.JwtAuthenticationFilter
 
@@ -23,6 +24,7 @@ class SecurityConfig(
     private val domainAuthorizationConfig: DomainAuthorizationConfig,
     @param:Qualifier("configure") private val corsConfigurationSource: CorsConfigurationSource,
     private val jwtProvider: JwtProvider,
+    private val customAuthenticationEntryPoint: CustomAuthenticationEntryPoint,
 ) {
     @Bean
     fun securityFilterChain(http: HttpSecurity): SecurityFilterChain {
@@ -33,6 +35,7 @@ class SecurityConfig(
             .formLogin(FormLoginConfigurer<*>::disable)
             .logout(LogoutConfigurer<*>::disable)
             .sessionManagement { it.sessionCreationPolicy(SessionCreationPolicy.STATELESS) }
+            .exceptionHandling { it.authenticationEntryPoint(customAuthenticationEntryPoint) }
             .addFilterBefore(JwtAuthenticationFilter(jwtProvider), UsernamePasswordAuthenticationFilter::class.java)
             .authorizeHttpRequests { domainAuthorizationConfig.configure(it) }
